+++ Sources/PresentationKit/Alerts/AlertableError.swift
@@ -2,7 +2,7 @@
 //  AlertableError.swift
 //  PresentationKit
 //
-//  Created by Daniel Saidi on 2025-06-23.
+//  Created by Daniel Saidi on 2025-06-19.
 //  Copyright © 2025-2026 Daniel Saidi. All rights reserved.
 //
 
@@ -19,58 +19,4 @@ public protocol AlertableError: Error {
 
     /// The alert message to display.
     var alertMessage: AlertMessage { get }
-}
-
-// MARK: - View Extensions
-
-public extension View {
-
-    /// Presents an alert when the context item is set.
-    ///
-    /// This modifier tries to cast the item to ``AlertError``
-    /// and uses its ``AlertError/alertMessage`` if possible.
-    /// If the cast fails, a default alert is shown with the
-    /// error's `localizedDescription` message.
-    func alert<Item: Error>(
-        for context: Binding<Presentation<Item>>
-    ) -> some View {
-        self.alert(
-            alertTitle(for: context.wrappedValue.item),
-            isPresented: Binding(
-                get: { context.wrappedValue.item != nil },
-                set: { if !$0 { context.wrappedValue.item = nil } }
-            ),
-            presenting: context.wrappedValue.item,
-            actions: { alertActions(for: $0) },
-            message: { alertMessage(for: $0) }
-        )
-    }
-}
-
-private extension View {
-
-    func alertTitle(for item: (any Error)?) -> String {
-        if let alertError = item as? any AlertableError {
-            return alertError.alertMessage.title
-        }
-        return "Error"
-    }
-
-    @ViewBuilder
-    func alertActions(for item: any Error) -> some View {
-        if let alertError = item as? any AlertableError {
-            alertError.alertMessage.actions()
-        } else {
-            Button("OK") {}
-        }
-    }
-
-    @ViewBuilder
-    func alertMessage(for item: any Error) -> some View {
-        if let alertError = item as? any AlertableError {
-            alertError.alertMessage.message()
-        } else {
-            Text(item.localizedDescription)
-        }
-    }
 }